+++ src/interface/variable.hpp
@@ -118,7 +118,9 @@ class CellVariable {
   inline std::string getAssociated() { return m_.getAssociated(); }
 
   KOKKOS_FORCEINLINE_FUNCTION
-  std::size_t GetUniqueID() const { return uid_; }
+  Uid_t GetUniqueID() const { return uid_; }
+
+  static Uid_t GetUniqueID(const std::string &var_label) { return get_uid_(var_label); }
 
   /// return information string
   std::string info();
@@ -165,9 +167,9 @@ class CellVariable {
   const std::array<int, 6> dims_, coarse_dims_;
 
   // Machinery for giving each variable a unique ID that is faster to
-  // evaluate than a string. uid is determined by order of variable
-  // creation, which is deterministic.
-  std::size_t uid_;
+  // evaluate than a string. Safe so long as the number of MPI ranks
+  // does not change while the code is running (restarts are fine).
+  Uid_t uid_;
   // This generator needs to be global so that different instances of
   // variable have the same unique ID.
   inline static UniqueIDGenerator<std::string> get_uid_;
@@ -188,6 +190,13 @@ class ParticleVariable {
   template <class... Args>
   KOKKOS_FORCEINLINE_FUNCTION auto &operator()(Args... args) {
     return data(std::forward<Args>(args)...);
+  }
+
+  auto GetHostMirrorAndCopy() { return data.GetHostMirrorAndCopy(); }
+
+  auto GetHostMirrorAndCopy(int n6, int n5, int n4, int n3, int n2) {
+    auto data_slice = Kokkos::subview(data, n6, n5, n4, n3, n2, Kokkos::ALL());
+    return data_slice.GetHostMirrorAndCopy();
   }
 
   KOKKOS_FORCEINLINE_FUNCTION
@@ -271,9 +280,11 @@ template <typename T>
 using MapToCellVars = std::map<std::string, std::shared_ptr<CellVariable<T>>>;
 
 template <typename T>
-using ParticleVariableVector = std::vector<std::shared_ptr<ParticleVariable<T>>>;
+using ParticleVarPtr = std::shared_ptr<ParticleVariable<T>>;
 template <typename T>
-using MapToParticle = std::map<std::string, std::shared_ptr<ParticleVariable<T>>>;
+using ParticleVariableVector = std::vector<ParticleVarPtr<T>>;
+template <typename T>
+using MapToParticle = std::map<std::string, ParticleVarPtr<T>>;
 template <typename T>
 using VariableSet = std::set<VarPtr<T>, VarComp<CellVariable<T>>>;
 template <typename T>